remove mixin_fab_tests.h
